Refactor and enhance logging in analysis and triage agents
- Improved logging statements for better clarity and consistency across the AnalysisAgent and TriageAgent classes.
- Updated method signatures to include additional parameters for enhanced functionality.
- Added regex validation for project_id and location in ServiceMonitorConfig.
- Removed the pattern_detector module and its associated test suite to streamline the codebase.

This commit aims to improve code maintainability and readability while ensuring robust logging for better debugging and monitoring.
