+++ gemini_sre_agent/analysis_agent.py
@@ -1,28 +1,39 @@
-import logging
 import json
-from typing import List, Dict, Any
-from pydantic import BaseModel, ValidationError
+import logging
+from typing import Any, Dict, List
+
 from google.cloud import aiplatform
+from pydantic import BaseModel, ValidationError
+from tenacity import (
+    retry,
+    retry_if_exception_type,
+    stop_after_attempt,
+    wait_exponential,
+)
 from vertexai.preview.generative_models import GenerativeModel
+
 from .triage_agent import TriagePacket
-from tenacity import retry, stop_after_attempt, wait_exponential, retry_if_exception_type
 
 logger = logging.getLogger(__name__)
 
+
 class RemediationPlan(BaseModel):
     """
-    Represents a detailed plan for remediating a service code issue, including root cause, 
+    Represents a detailed plan for remediating a service code issue, including root cause,
     proposed fix, and code patch.
     """
+
     root_cause_analysis: str
     proposed_fix: str
     code_patch: str
+
 
 class AnalysisAgent:
     """
     A class responsible for performing deep analysis of issues using a Gemini model
     to generate remediation plans.
     """
+
     def __init__(self, project_id: str, location: str, analysis_model: str):
         """
         Initializes the AnalysisAgent with GCP project, location, and the analysis model to use.
@@ -37,15 +48,21 @@ class AnalysisAgent:
         self.analysis_model: str = analysis_model
         aiplatform.init(project=project_id, location=location)
         self.model: GenerativeModel = GenerativeModel(analysis_model)
-        logger.info(f"[ANALYSIS] AnalysisAgent initialized with model: {analysis_model} in {location} for project: {project_id}")
+        logger.info(
+            f"[ANALYSIS] AnalysisAgent initialized with model: {analysis_model} in {location} for project: {project_id}"
+        )
 
     @retry(
         stop=stop_after_attempt(3),
         wait=wait_exponential(multiplier=1, min=4, max=10),
-        retry=retry_if_exception_type((RuntimeError, ValueError, json.JSONDecodeError))
+        retry=retry_if_exception_type((RuntimeError, ValueError, json.JSONDecodeError)),
     )
     def analyze_issue(
-        self, triage_packet: TriagePacket, historical_logs: List[str], configs: Dict[str, str], flow_id: str
+        self,
+        triage_packet: TriagePacket,
+        historical_logs: List[str],
+        configs: Dict[str, str],
+        flow_id: str,
     ) -> RemediationPlan:
         """
         Analyzes an issue based on a triage packet, historical logs, and configurations
@@ -60,8 +77,10 @@ class AnalysisAgent:
         Returns:
             RemediationPlan: A structured plan for remediating the issue.
         """
-        logger.info(f"[ANALYSIS] Analyzing issue: flow_id={flow_id}, issue_id={triage_packet.issue_id}, historical_logs={len(historical_logs)}, configs={len(configs)}")
-        
+        logger.info(
+            f"[ANALYSIS] Analyzing issue: flow_id={flow_id}, issue_id={triage_packet.issue_id}, historical_logs={len(historical_logs)}, configs={len(configs)}"
+        )
+
         # Construct the prompt for the Gemini model
         prompt_template: str = """
         You are an expert SRE Analysis Agent. Your task is to perform a deep root cause analysis of the provided issue,
@@ -86,36 +105,48 @@ class AnalysisAgent:
 
         Provide only the JSON response.
         """
-        
+
         prompt: str = prompt_template.format(
             triage_packet_json=triage_packet.model_dump_json(),
             historical_logs_str=json.dumps(historical_logs, indent=2),
-            configs_str=json.dumps(configs, indent=2)
+            configs_str=json.dumps(configs, indent=2),
         )
-        logger.debug(f"[ANALYSIS] Prompt for analysis model: flow_id={flow_id}, issue_id={triage_packet.issue_id}, prompt={prompt[:500]}...")
+        logger.debug(
+            f"[ANALYSIS] Prompt for analysis model: flow_id={flow_id}, issue_id={triage_packet.issue_id}, prompt={prompt[:500]}..."
+        )
 
-        json_response_str: str = "" # Initialize json_response_str
+        json_response_str: str = ""  # Initialize json_response_str
 
         try:
             # Call the Gemini model (synchronous)
-            response = self.model.generate_content(prompt) # Removed await
-            
+            response = self.model.generate_content(prompt)  # Removed await
+
             # Extract and parse the JSON response
             json_response_str = response.text.strip()
-            logger.debug(f"[ANALYSIS] Raw model response: flow_id={flow_id}, issue_id={triage_packet.issue_id}, response={json_response_str[:500]}...")
+            logger.debug(
+                f"[ANALYSIS] Raw model response: flow_id={flow_id}, issue_id={triage_packet.issue_id}, response={json_response_str[:500]}..."
+            )
 
             remediation_data: Dict[str, Any] = json.loads(json_response_str)
             remediation_plan: RemediationPlan = RemediationPlan(**remediation_data)
-            
-            logger.info(f"[ANALYSIS] Analysis complete: flow_id={flow_id}, issue_id={triage_packet.issue_id}")
+
+            logger.info(
+                f"[ANALYSIS] Analysis complete: flow_id={flow_id}, issue_id={triage_packet.issue_id}"
+            )
             return remediation_plan
 
         except ValidationError as e:
-            logger.error(f"[ERROR_HANDLING] Failed to validate RemediationPlan schema from model response: flow_id={flow_id}, issue_id={triage_packet.issue_id}, error={e}")
+            logger.error(
+                f"[ERROR_HANDLING] Failed to validate RemediationPlan schema from model response: flow_id={flow_id}, issue_id={triage_packet.issue_id}, error={e}"
+            )
             raise ValueError(f"Invalid model response schema: {e}") from e
         except json.JSONDecodeError as e:
-            logger.error(f"[ERROR_HANDLING] Failed to decode JSON from model response: flow_id={flow_id}, issue_id={triage_packet.issue_id}, error={e}, response={json_response_str}")
+            logger.error(
+                f"[ERROR_HANDLING] Failed to decode JSON from model response: flow_id={flow_id}, issue_id={triage_packet.issue_id}, error={e}, response={json_response_str}"
+            )
             raise ValueError(f"Malformed JSON response from model: {e}") from e
         except Exception as e:
-            logger.error(f"[ERROR_HANDLING] Error calling Gemini Analysis model: flow_id={flow_id}, issue_id={triage_packet.issue_id}, error={e}")
+            logger.error(
+                f"[ERROR_HANDLING] Error calling Gemini Analysis model: flow_id={flow_id}, issue_id={triage_packet.issue_id}, error={e}"
+            )
             raise RuntimeError(f"Gemini Analysis model call failed: {e}") from e
