+++ gemini_sre_agent/pattern_detector/__init__.py
@@ -1,0 +1,40 @@
+"""
+Pattern Detection System
+"""
+
+from .baseline_tracker import BaselineTracker
+from .confidence_scorer import ConfidenceScorer
+from .models import (
+    ConfidenceFactors,
+    ConfidenceRule,
+    ConfidenceScore,
+    LogEntry,
+    PatternMatch,
+    PatternType,
+    ThresholdConfig,
+    ThresholdResult,
+    ThresholdType,
+    TimeWindow,
+)
+from .pattern_classifier import PatternClassifier
+from .threshold_evaluator import ThresholdEvaluator
+from .time_window_accumulator import LogAccumulator, WindowManager
+
+__all__ = [
+    "BaselineTracker",
+    "ConfidenceFactors",
+    "ConfidenceRule",
+    "ConfidenceScore",
+    "ConfidenceScorer",
+    "LogAccumulator",
+    "LogEntry",
+    "PatternClassifier",
+    "PatternMatch",
+    "PatternType",
+    "ThresholdConfig",
+    "ThresholdEvaluator",
+    "ThresholdResult",
+    "ThresholdType",
+    "TimeWindow",
+    "WindowManager",
+]
