+++ gemini_sre_agent/triage_agent.py
@@ -1,17 +1,25 @@
-import logging
 import json
-from typing import List, Dict, Any
-from pydantic import BaseModel, ValidationError
+import logging
+from typing import Any, Dict, List
+
 from google.cloud import aiplatform
+from pydantic import BaseModel, ValidationError
+from tenacity import (
+    retry,
+    retry_if_exception_type,
+    stop_after_attempt,
+    wait_exponential,
+)
 from vertexai.preview.generative_models import GenerativeModel
-from tenacity import retry, stop_after_attempt, wait_exponential, retry_if_exception_type
 
 logger = logging.getLogger(__name__)
+
 
 class TriagePacket(BaseModel):
     """
     Represents a packet of information generated by the Triage Agent after analyzing logs.
     """
+
     issue_id: str
     initial_timestamp: str
     detected_pattern: str
@@ -20,10 +28,12 @@ class TriagePacket(BaseModel):
     sample_log_entries: List[str]
     natural_language_summary: str
 
+
 class TriageAgent:
     """
     A class responsible for triaging log entries using a Gemini model.
     """
+
     def __init__(self, project_id: str, location: str, triage_model: str):
         """
         Initializes the TriageAgent with GCP project, location, and the triage model to use.
@@ -38,12 +48,14 @@ class TriageAgent:
         self.triage_model: str = triage_model
         aiplatform.init(project=project_id, location=location)
         self.model: GenerativeModel = GenerativeModel(triage_model)
-        logger.info(f"[TRIAGE] TriageAgent initialized with model: {triage_model} in {location} for project: {project_id}")
+        logger.info(
+            f"[TRIAGE] TriageAgent initialized with model: {triage_model} in {location} for project: {project_id}"
+        )
 
     @retry(
         stop=stop_after_attempt(3),
         wait=wait_exponential(multiplier=1, min=4, max=10),
-        retry=retry_if_exception_type((RuntimeError, ValueError, json.JSONDecodeError))
+        retry=retry_if_exception_type((RuntimeError, ValueError, json.JSONDecodeError)),
     )
     async def analyze_logs(self, logs: List[str], flow_id: str) -> TriagePacket:
         """
@@ -56,8 +68,10 @@ class TriageAgent:
         Returns:
             TriagePacket: A structured packet containing triage information.
         """
-        logger.info(f"[TRIAGE] Analyzing {len(logs)} log entries for triage: flow_id={flow_id}")
-        
+        logger.info(
+            f"[TRIAGE] Analyzing {len(logs)} log entries for triage: flow_id={flow_id}"
+        )
+
         # Construct the prompt for the Gemini model
         prompt_template: str = """
         You are an expert SRE Triage Agent. Your task is to analyze the provided log entries, identify any critical issues, and summarize them into a structured JSON format.
@@ -78,31 +92,43 @@ class TriageAgent:
         Provide only the JSON response.
         """
         prompt: str = prompt_template.format(log_entries="\n".join(logs))
-        logger.debug(f"[TRIAGE] Prompt for triage model: flow_id={flow_id}, prompt={prompt[:500]}...")
+        logger.debug(
+            f"[TRIAGE] Prompt for triage model: flow_id={flow_id}, prompt={prompt[:500]}..."
+        )
 
-        json_response_str: str = "" # Initialize json_response_str
+        json_response_str: str = ""  # Initialize json_response_str
 
         try:
             # Call the Gemini model (synchronous)
             response = self.model.generate_content(prompt)
-            
+
             # Extract and parse the JSON response
             json_response_str = response.text.strip()
-            logger.debug(f"[TRIAGE] Raw model response: flow_id={flow_id}, response={json_response_str[:500]}...")
+            logger.debug(
+                f"[TRIAGE] Raw model response: flow_id={flow_id}, response={json_response_str[:500]}..."
+            )
 
             # Attempt to parse the JSON response
             triage_data: Dict[str, Any] = json.loads(json_response_str)
             triage_packet: TriagePacket = TriagePacket(**triage_data)
-            
-            logger.info(f"[TRIAGE] Triage analysis complete: flow_id={flow_id}, issue_id={triage_packet.issue_id}, severity={triage_packet.preliminary_severity_score}")
+
+            logger.info(
+                f"[TRIAGE] Triage analysis complete: flow_id={flow_id}, issue_id={triage_packet.issue_id}, severity={triage_packet.preliminary_severity_score}"
+            )
             return triage_packet
 
         except ValidationError as e:
-            logger.error(f"[ERROR_HANDLING] Failed to validate TriagePacket schema from model response: flow_id={flow_id}, error={e}")
+            logger.error(
+                f"[ERROR_HANDLING] Failed to validate TriagePacket schema from model response: flow_id={flow_id}, error={e}"
+            )
             raise ValueError(f"Invalid model response schema: {e}") from e
         except json.JSONDecodeError as e:
-            logger.error(f"[ERROR_HANDLING] Failed to decode JSON from model response: flow_id={flow_id}, error={e}, response={json_response_str}")
+            logger.error(
+                f"[ERROR_HANDLING] Failed to decode JSON from model response: flow_id={flow_id}, error={e}, response={json_response_str}"
+            )
             raise ValueError(f"Malformed JSON response from model: {e}") from e
         except Exception as e:
-            logger.error(f"[ERROR_HANDLING] Error calling Gemini Triage model: flow_id={flow_id}, error={e}")
+            logger.error(
+                f"[ERROR_HANDLING] Error calling Gemini Triage model: flow_id={flow_id}, error={e}"
+            )
             raise RuntimeError(f"Gemini Triage model call failed: {e}") from e
