--- tests/test_pattern_detector.py
@@ -1,2304 +1,0 @@
-"""
-Test suite for Pattern Detection System - Layer 1: Time-Window Accumulation
-
-This test suite validates the time-window accumulation functionality including:
-- LogEntry creation and parsing
-- TimeWindow management and log acceptance
-- LogAccumulator with sliding windows
-- WindowManager dual-window coordination
-"""
-
-import pytest
-import asyncio
-import json
-from datetime import datetime, timedelta, timezone
-from unittest.mock import MagicMock, AsyncMock, patch
-from typing import List, Dict, Any
-
-from gemini_sre_agent.pattern_detector import (
-    LogEntry,
-    TimeWindow,
-    LogAccumulator,
-    WindowManager,
-    ThresholdType,
-    ThresholdConfig,
-    ThresholdResult,
-    BaselineTracker,
-    ThresholdEvaluator,
-    PatternType,
-    PatternMatch,
-    PatternClassifier,
-    ConfidenceFactors,
-    ConfidenceRule,
-    ConfidenceScore,
-    ConfidenceScorer
-)
-
-
-class TestLogEntry:
-    """Test LogEntry creation and parsing functionality."""
-    
-    def test_log_entry_basic_creation(self):
-        """Test basic LogEntry creation with all fields provided."""
-        timestamp = datetime.now(timezone.utc)
-        raw_data = {"severity": "ERROR", "textPayload": "Test error"}
-        
-        log_entry = LogEntry(
-            insert_id="test-123",
-            timestamp=timestamp,
-            severity="ERROR",
-            service_name="test-service",
-            error_message="Test error",
-            raw_data=raw_data
-        )
-        
-        assert log_entry.insert_id == "test-123"
-        assert log_entry.timestamp == timestamp
-        assert log_entry.severity == "ERROR"
-        assert log_entry.service_name == "test-service"
-        assert log_entry.error_message == "Test error"
-        assert log_entry.raw_data == raw_data
-    
-    def test_log_entry_timestamp_extraction(self):
-        """Test automatic timestamp extraction from raw data."""
-        raw_data = {
-            "timestamp": "2025-01-27T10:00:00Z",
-            "severity": "ERROR",
-            "textPayload": "Test error"
-        }
-        
-        log_entry = LogEntry(
-            insert_id="test-123",
-            raw_data=raw_data
-        )
-        
-        assert log_entry.timestamp.year == 2025
-        assert log_entry.timestamp.month == 1
-        assert log_entry.timestamp.day == 27
-        assert log_entry.timestamp.hour == 10
-    
-    def test_log_entry_service_name_extraction(self):
-        """Test automatic service name extraction from resource labels."""
-        raw_data = {
-            "timestamp": "2025-01-27T10:00:00Z",
-            "resource": {
-                "type": "cloud_run_revision",
-                "labels": {
-                    "service_name": "billing-service",
-                    "revision_name": "billing-service-001"
-                }
-            },
-            "severity": "ERROR"
-        }
-        
-        log_entry = LogEntry(
-            insert_id="test-123",
-            raw_data=raw_data
-        )
-        
-        assert log_entry.service_name == "billing-service"
-    
-    def test_log_entry_function_name_extraction(self):
-        """Test service name extraction from function_name label."""
-        raw_data = {
-            "timestamp": "2025-01-27T10:00:00Z",
-            "resource": {
-                "type": "cloud_function",
-                "labels": {
-                    "function_name": "payment-processor"
-                }
-            },
-            "severity": "ERROR"
-        }
-        
-        log_entry = LogEntry(
-            insert_id="test-123",
-            raw_data=raw_data
-        )
-        
-        assert log_entry.service_name == "payment-processor"
-    
-    def test_log_entry_error_message_extraction(self):
-        """Test automatic error message extraction from textPayload."""
-        raw_data = {
-            "timestamp": "2025-01-27T10:00:00Z",
-            "textPayload": "Database connection failed: timeout after 30s",
-            "severity": "ERROR"
-        }
-        
-        log_entry = LogEntry(
-            insert_id="test-123",
-            raw_data=raw_data
-        )
-        
-        assert log_entry.error_message == "Database connection failed: timeout after 30s"
-    
-    def test_log_entry_invalid_timestamp_fallback(self):
-        """Test fallback to current time for invalid timestamps."""
-        raw_data = {
-            "timestamp": "invalid-timestamp",
-            "severity": "ERROR"
-        }
-        
-        before_creation = datetime.now(timezone.utc)
-        log_entry = LogEntry(
-            insert_id="test-123",
-            raw_data=raw_data
-        )
-        after_creation = datetime.now(timezone.utc)
-        
-        # Should fallback to current time
-        assert before_creation <= log_entry.timestamp <= after_creation
-
-
-class TestTimeWindow:
-    """Test TimeWindow functionality."""
-    
-    def test_time_window_creation(self):
-        """Test basic TimeWindow creation."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        assert window.start_time == start_time
-        assert window.duration_minutes == 5
-        assert window.end_time == datetime(2025, 1, 27, 10, 5, 0)
-        assert len(window.logs) == 0
-    
-    def test_time_window_is_active(self):
-        """Test window active status checking."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        # Should be active before end time
-        assert window.is_active(datetime(2025, 1, 27, 10, 3, 0)) is True
-        
-        # Should not be active at end time
-        assert window.is_active(datetime(2025, 1, 27, 10, 5, 0)) is False
-        
-        # Should not be active after end time
-        assert window.is_active(datetime(2025, 1, 27, 10, 6, 0)) is False
-    
-    def test_time_window_is_expired(self):
-        """Test window expiration checking."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        # Should not be expired before end time
-        assert window.is_expired(datetime(2025, 1, 27, 10, 3, 0)) is False
-        
-        # Should be expired at end time
-        assert window.is_expired(datetime(2025, 1, 27, 10, 5, 0)) is True
-        
-        # Should be expired after end time
-        assert window.is_expired(datetime(2025, 1, 27, 10, 6, 0)) is True
-    
-    def test_time_window_accepts_log(self):
-        """Test log acceptance logic."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        # Create test log entries
-        log_before = LogEntry(
-            insert_id="before",
-            timestamp=datetime(2025, 1, 27, 9, 59, 0),
-            severity="ERROR",
-            raw_data={}
-        )
-        
-        log_within = LogEntry(
-            insert_id="within",
-            timestamp=datetime(2025, 1, 27, 10, 3, 0),
-            severity="ERROR",
-            raw_data={}
-        )
-        
-        log_after = LogEntry(
-            insert_id="after",
-            timestamp=datetime(2025, 1, 27, 10, 6, 0),
-            severity="ERROR",
-            raw_data={}
-        )
-        
-        assert window.accepts_log(log_before) is False
-        assert window.accepts_log(log_within) is True
-        assert window.accepts_log(log_after) is False
-    
-    def test_time_window_add_log(self):
-        """Test adding logs to window."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        # Create test logs
-        valid_log = LogEntry(
-            insert_id="valid",
-            timestamp=datetime(2025, 1, 27, 10, 3, 0),
-            severity="ERROR",
-            raw_data={}
-        )
-        
-        invalid_log = LogEntry(
-            insert_id="invalid",
-            timestamp=datetime(2025, 1, 27, 10, 6, 0),
-            severity="ERROR",
-            raw_data={}
-        )
-        
-        # Valid log should be added
-        assert window.add_log(valid_log) is True
-        assert len(window.logs) == 1
-        assert window.logs[0] == valid_log
-        
-        # Invalid log should be rejected
-        assert window.add_log(invalid_log) is False
-        assert len(window.logs) == 1
-    
-    def test_time_window_get_error_logs(self):
-        """Test filtering for error-level logs."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        # Add logs of different severities
-        logs = [
-            LogEntry(insert_id="info", timestamp=start_time, severity="INFO", raw_data={}),
-            LogEntry(insert_id="error", timestamp=start_time, severity="ERROR", raw_data={}),
-            LogEntry(insert_id="critical", timestamp=start_time, severity="CRITICAL", raw_data={}),
-            LogEntry(insert_id="warning", timestamp=start_time, severity="WARNING", raw_data={})
-        ]
-        
-        for log in logs:
-            window.add_log(log)
-        
-        error_logs = window.get_error_logs()
-        assert len(error_logs) == 2
-        assert error_logs[0].severity == "ERROR"
-        assert error_logs[1].severity == "CRITICAL"
-    
-    def test_time_window_get_service_groups(self):
-        """Test grouping logs by service."""
-        start_time = datetime(2025, 1, 27, 10, 0, 0)
-        window = TimeWindow(start_time=start_time, duration_minutes=5)
-        
-        # Add logs from different services
-        logs = [
-            LogEntry(insert_id="1", timestamp=start_time, service_name="service-a", severity="ERROR", raw_data={}),
-            LogEntry(insert_id="2", timestamp=start_time, service_name="service-a", severity="INFO", raw_data={}),
-            LogEntry(insert_id="3", timestamp=start_time, service_name="service-b", severity="ERROR", raw_data={}),
-            LogEntry(insert_id="4", timestamp=start_time, service_name=None, severity="WARN", raw_data={})
-        ]
-        
-        for log in logs:
-            window.add_log(log)
-        
-        groups = window.get_service_groups()
-        assert len(groups) == 3
-        assert len(groups["service-a"]) == 2
-        assert len(groups["service-b"]) == 1
-        assert len(groups["unknown"]) == 1
-
-
-class TestLogAccumulator:
-    """Test LogAccumulator functionality."""
-    
-    @pytest.fixture
-    def mock_callback(self):
-        """Create mock callback for testing."""
-        return MagicMock()
-    
-    def test_log_accumulator_initialization(self, mock_callback):
-        """Test LogAccumulator initialization."""
-        accumulator = LogAccumulator(
-            window_duration_minutes=5,
-            max_windows=10,
-            on_window_ready=mock_callback
-        )
-        
-        assert accumulator.window_duration_minutes == 5
-        assert accumulator.max_windows == 10
-        assert accumulator.on_window_ready == mock_callback
-        assert len(accumulator.windows) == 0
-    
-    def test_log_accumulator_round_to_window_start(self, mock_callback):
-        """Test timestamp rounding to window boundaries."""
-        accumulator = LogAccumulator(
-            window_duration_minutes=5,
-            on_window_ready=mock_callback
-        )
-        
-        # Test various timestamps
-        test_cases = [
-            (datetime(2025, 1, 27, 10, 3, 30), datetime(2025, 1, 27, 10, 0, 0)),
-            (datetime(2025, 1, 27, 10, 7, 15), datetime(2025, 1, 27, 10, 5, 0)),
-            (datetime(2025, 1, 27, 10, 12, 45), datetime(2025, 1, 27, 10, 10, 0)),
-        ]
-        
-        for input_time, expected_start in test_cases:
-            result = accumulator._round_to_window_start(input_time)
-            assert result == expected_start
-    
-    def test_log_accumulator_add_log(self, mock_callback):
-        """Test adding logs to accumulator."""
-        accumulator = LogAccumulator(
-            window_duration_minutes=5,
-            on_window_ready=mock_callback
-        )
-        
-        # Add a log
-        log_data = {
-            "insertId": "test-123",
-            "timestamp": "2025-01-27T10:03:00Z",
-            "severity": "ERROR",
-            "textPayload": "Test error"
-        }
-        
-        accumulator.add_log(log_data)
-        
-        # Should create one window
-        assert len(accumulator.windows) == 1
-        
-        # Window should contain the log
-        window = list(accumulator.windows.values())[0]
-        assert len(window.logs) == 1
-        assert window.logs[0].insert_id == "test-123"
-    
-    def test_log_accumulator_multiple_windows(self, mock_callback):
-        """Test creation of multiple windows for different time ranges."""
-        accumulator = LogAccumulator(
-            window_duration_minutes=5,
-            on_window_ready=mock_callback
-        )
-        
-        # Add logs in different 5-minute windows
-        log_data_1 = {
-            "insertId": "test-1",
-            "timestamp": "2025-01-27T10:03:00Z",  # 10:00-10:05 window
-            "severity": "ERROR"
-        }
-        
-        log_data_2 = {
-            "insertId": "test-2",
-            "timestamp": "2025-01-27T10:07:00Z",  # 10:05-10:10 window
-            "severity": "ERROR"
-        }
-        
-        accumulator.add_log(log_data_1)
-        accumulator.add_log(log_data_2)
-        
-        # Should create two windows
-        assert len(accumulator.windows) == 2
-    
-    def test_log_accumulator_window_eviction(self, mock_callback):
-        """Test window eviction when max_windows is exceeded."""
-        accumulator = LogAccumulator(
-            window_duration_minutes=5,
-            max_windows=2,
-            on_window_ready=mock_callback
-        )
-        
-        # Add logs to create 3 windows (should trigger eviction)
-        timestamps = [
-            "2025-01-27T10:03:00Z",  # Window 1: 10:00-10:05
-            "2025-01-27T10:07:00Z",  # Window 2: 10:05-10:10  
-            "2025-01-27T10:13:00Z"   # Window 3: 10:10-10:15 (should evict window 1)
-        ]
-        
-        for i, timestamp in enumerate(timestamps):
-            log_data = {
-                "insertId": f"test-{i}",
-                "timestamp": timestamp,
-                "severity": "ERROR"
-            }
-            accumulator.add_log(log_data)
-        
-        # Should only have max_windows (2) windows
-        assert len(accumulator.windows) == 2
-        
-        # Callback should have been called for evicted window
-        mock_callback.assert_called_once()
-    
-    @pytest.mark.asyncio
-    async def test_log_accumulator_start_stop(self, mock_callback):
-        """Test starting and stopping the accumulator."""
-        accumulator = LogAccumulator(on_window_ready=mock_callback)
-        
-        # Start should create cleanup task
-        accumulator.start()
-        assert accumulator._cleanup_task is not None
-        assert not accumulator._cleanup_task.done()
-        
-        # Stop should cancel task and process remaining windows
-        await accumulator.stop()
-        assert accumulator._shutdown is True
-
-
-class TestWindowManager:
-    """Test WindowManager functionality."""
-    
-    @pytest.fixture
-    def mock_pattern_callback(self):
-        """Create mock pattern callback for testing."""
-        return MagicMock()
-    
-    def test_window_manager_initialization(self, mock_pattern_callback):
-        """Test WindowManager initialization."""
-        manager = WindowManager(
-            fast_window_minutes=5,
-            trend_window_minutes=15,
-            max_windows=20,
-            pattern_callback=mock_pattern_callback
-        )
-        
-        assert manager.fast_window_minutes == 5
-        assert manager.trend_window_minutes == 15
-        assert manager.pattern_callback == mock_pattern_callback
-        assert manager.fast_accumulator is not None
-        assert manager.trend_accumulator is not None
-    
-    @pytest.mark.asyncio
-    async def test_window_manager_start_stop(self, mock_pattern_callback):
-        """Test starting and stopping the window manager."""
-        manager = WindowManager(pattern_callback=mock_pattern_callback)
-        
-        # Mock the accumulator methods
-        manager.fast_accumulator.start = MagicMock()
-        manager.trend_accumulator.start = MagicMock()
-        manager.fast_accumulator.stop = AsyncMock()
-        manager.trend_accumulator.stop = AsyncMock()
-        
-        # Start should call start on both accumulators
-        manager.start()
-        manager.fast_accumulator.start.assert_called_once()
-        manager.trend_accumulator.start.assert_called_once()
-        
-        # Stop should call stop on both accumulators
-        await manager.stop()
-        manager.fast_accumulator.stop.assert_called_once()
-        manager.trend_accumulator.stop.assert_called_once()
-    
-    def test_window_manager_add_log(self, mock_pattern_callback):
-        """Test adding logs to both accumulators."""
-        manager = WindowManager(pattern_callback=mock_pattern_callback)
-        
-        # Mock the add_log methods
-        manager.fast_accumulator.add_log = MagicMock()
-        manager.trend_accumulator.add_log = MagicMock()
-        
-        log_data = {
-            "insertId": "test-123",
-            "timestamp": "2025-01-27T10:03:00Z",
-            "severity": "ERROR"
-        }
-        
-        manager.add_log(log_data)
-        
-        # Should call add_log on both accumulators
-        manager.fast_accumulator.add_log.assert_called_once_with(log_data)
-        manager.trend_accumulator.add_log.assert_called_once_with(log_data)
-    
-    def test_window_manager_fast_window_callback(self, mock_pattern_callback):
-        """Test fast window completion callback."""
-        manager = WindowManager(pattern_callback=mock_pattern_callback)
-        
-        # Create mock window
-        mock_window = MagicMock()
-        mock_window.start_time = datetime(2025, 1, 27, 10, 0, 0)
-        mock_window.logs = ["log1", "log2"]
-        mock_window.get_error_logs.return_value = ["error1"]
-        
-        # Call the fast window callback
-        manager._on_fast_window_ready(mock_window)
-        
-        # Should call the pattern callback
-        mock_pattern_callback.assert_called_once_with(mock_window)
-    
-    def test_window_manager_trend_window_callback(self, mock_pattern_callback):
-        """Test trend window completion callback."""
-        manager = WindowManager(pattern_callback=mock_pattern_callback)
-        
-        # Create mock window
-        mock_window = MagicMock()
-        mock_window.start_time = datetime(2025, 1, 27, 10, 0, 0)
-        mock_window.logs = ["log1", "log2", "log3"]
-        mock_window.get_error_logs.return_value = ["error1", "error2"]
-        
-        # Call the trend window callback
-        manager._on_trend_window_ready(mock_window)
-        
-        # Should call the pattern callback
-        mock_pattern_callback.assert_called_once_with(mock_window)
-    
-    def test_window_manager_callback_error_handling(self, mock_pattern_callback):
-        """Test error handling in window callbacks."""
-        # Make callback raise an exception
-        mock_pattern_callback.side_effect = RuntimeError("Test error")
-        
-        manager = WindowManager(pattern_callback=mock_pattern_callback)
-        
-        mock_window = MagicMock()
-        mock_window.start_time = datetime(2025, 1, 27, 10, 0, 0)
-        mock_window.logs = []
-        mock_window.get_error_logs.return_value = []
-        
-        # Should not raise exception (error should be logged)
-        with patch('gemini_sre_agent.pattern_detector.logger') as mock_logger:
-            manager._on_fast_window_ready(mock_window)
-            manager._on_trend_window_ready(mock_window)
-            
-            # Should log errors
-            assert mock_logger.error.call_count == 2
-
-
-@pytest.mark.integration
-class TestPatternDetectorIntegration:
-    """Integration tests for the complete pattern detection system."""
-    
-    @pytest.mark.asyncio
-    async def test_end_to_end_pattern_detection(self):
-        """Test complete flow from log ingestion to pattern detection."""
-        received_windows = []
-        
-        def pattern_callback(window):
-            received_windows.append(window)
-        
-        # Create manager with short windows for testing
-        manager = WindowManager(
-            fast_window_minutes=1,  # Very short for testing
-            trend_window_minutes=2,
-            pattern_callback=pattern_callback
-        )
-        
-        manager.start()
-        
-        try:
-            # Add logs over time
-            base_time = datetime.now(timezone.utc)
-            logs = [
-                {
-                    "insertId": f"test-{i}",
-                    "timestamp": (base_time + timedelta(seconds=i*10)).isoformat() + "Z",
-                    "severity": "ERROR" if i % 2 == 0 else "INFO",
-                    "textPayload": f"Test message {i}",
-                    "resource": {
-                        "labels": {"service_name": f"service-{i % 2}"}
-                    }
-                }
-                for i in range(6)
-            ]
-            
-            for log_data in logs:
-                manager.add_log(log_data)
-            
-            # Wait a bit for window processing
-            await asyncio.sleep(0.1)
-            
-            # Verify logs were distributed correctly
-            assert len(manager.fast_accumulator.windows) > 0
-            assert len(manager.trend_accumulator.windows) > 0
-            
-        finally:
-            await manager.stop()
-    
-    def test_realistic_log_data_processing(self):
-        """Test with realistic GCP log data structure."""
-        received_windows = []
-        
-        def pattern_callback(window):
-            received_windows.append(window)
-        
-        manager = WindowManager(
-            fast_window_minutes=5,
-            pattern_callback=pattern_callback
-        )
-        
-        # Realistic GCP Cloud Run log
-        realistic_log = {
-            "insertId": "1234567890abcdef",
-            "logName": "projects/my-project/logs/run.googleapis.com%2Frequest",
-            "receiveTimestamp": "2025-01-27T10:03:45.123456Z",
-            "resource": {
-                "type": "cloud_run_revision",
-                "labels": {
-                    "service_name": "billing-service",
-                    "revision_name": "billing-service-00001-abc",
-                    "location": "us-central1"
-                }
-            },
-            "severity": "ERROR",
-            "textPayload": "Database connection failed: java.sql.SQLException: Connection refused",
-            "timestamp": "2025-01-27T10:03:44.987654Z",
-            "trace": "projects/my-project/traces/abc123def456"
-        }
-        
-        manager.add_log(realistic_log)
-        
-        # Verify log was processed correctly
-        assert len(manager.fast_accumulator.windows) == 1
-        window = list(manager.fast_accumulator.windows.values())[0]
-        assert len(window.logs) == 1
-        
-        log_entry = window.logs[0]
-        assert log_entry.service_name == "billing-service"
-        assert log_entry.severity == "ERROR"
-        assert log_entry.error_message is not None
-        assert "Database connection failed" in log_entry.error_message
-
-
-# ==========================================
-# Layer 2 Tests: Smart Thresholds
-# ==========================================
-
-
-class TestThresholdConfig:
-    """Test ThresholdConfig functionality."""
-    
-    def test_threshold_config_creation(self):
-        """Test basic ThresholdConfig creation."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.ERROR_FREQUENCY,
-            min_value=5.0,
-            min_error_count=3
-        )
-        
-        assert config.threshold_type == ThresholdType.ERROR_FREQUENCY
-        assert config.min_value == 5.0
-        assert config.min_error_count == 3
-        assert config.min_rate_increase == 10.0  # Default value
-    
-    def test_threshold_config_defaults(self):
-        """Test ThresholdConfig default values."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.ERROR_RATE,
-            min_value=15.0
-        )
-        
-        assert config.min_error_count == 3
-        assert config.min_rate_increase == 10.0
-        assert config.baseline_window_count == 12
-        assert config.min_affected_services == 2
-        assert "CRITICAL" in config.severity_weights
-        assert config.severity_weights["CRITICAL"] == 10.0
-
-
-class TestBaselineTracker:
-    """Test BaselineTracker functionality."""
-    
-    @pytest.fixture
-    def sample_windows(self):
-        """Create sample time windows for testing."""
-        base_time = datetime(2025, 1, 27, 10, 0, 0)
-        
-        windows = []
-        for i in range(5):
-            window = TimeWindow(
-                start_time=base_time + timedelta(minutes=i*5),
-                duration_minutes=5
-            )
-            
-            # Add logs with different error rates
-            for j in range(10):
-                severity = "ERROR" if j < i+1 else "INFO"  # Increasing error rate
-                log = LogEntry(
-                    insert_id=f"log-{i}-{j}",
-                    timestamp=window.start_time + timedelta(seconds=j*30),  # 30 second intervals within 5-minute window
-                    severity=severity,
-                    service_name=f"service-{j % 3}",
-                    raw_data={"severity": severity}
-                )
-                window.add_log(log)
-            
-            windows.append(window)
-        
-        return windows
-    
-    def test_baseline_tracker_initialization(self):
-        """Test BaselineTracker initialization."""
-        tracker = BaselineTracker(max_history=50)
-        
-        assert tracker.max_history == 50
-        assert len(tracker.global_baseline) == 0
-        assert len(tracker.service_baselines) == 0
-    
-    def test_baseline_tracker_update(self, sample_windows):
-        """Test updating baseline with window data."""
-        tracker = BaselineTracker()
-        
-        # Update with first window
-        tracker.update_baseline(sample_windows[0])
-        
-        assert len(tracker.global_baseline) == 1
-        assert tracker.global_baseline[0] == 10.0  # 1 error out of 10 logs
-        assert len(tracker.service_baselines) == 3  # 3 services
-    
-    def test_baseline_tracker_history_limit(self, sample_windows):
-        """Test baseline history limit enforcement."""
-        tracker = BaselineTracker(max_history=3)
-        
-        # Add more windows than limit
-        for window in sample_windows:
-            tracker.update_baseline(window)
-        
-        # Should only keep max_history windows
-        assert len(tracker.global_baseline) == 3
-        
-        # Should keep the most recent windows
-        expected_rates = [30.0, 40.0, 50.0]  # Error rates from last 3 windows
-        assert tracker.global_baseline == expected_rates
-    
-    def test_baseline_tracker_get_global_baseline(self, sample_windows):
-        """Test getting global baseline average."""
-        tracker = BaselineTracker()
-        
-        # Add sample windows
-        for window in sample_windows[:3]:
-            tracker.update_baseline(window)
-        
-        # Get baseline for last 2 windows
-        baseline = tracker.get_global_baseline(2)
-        expected = (20.0 + 30.0) / 2  # Average of last 2 rates
-        assert baseline == expected
-    
-    def test_baseline_tracker_get_service_baseline(self, sample_windows):
-        """Test getting service-specific baseline."""
-        tracker = BaselineTracker()
-        
-        # Add sample windows
-        for window in sample_windows[:3]:
-            tracker.update_baseline(window)
-        
-        # Get service baseline
-        service_baseline = tracker.get_service_baseline("service-0", 2)
-        assert service_baseline >= 0.0  # Should have some baseline
-
-
-class TestThresholdEvaluator:
-    """Test ThresholdEvaluator functionality."""
-    
-    @pytest.fixture
-    def sample_window_with_errors(self):
-        """Create a sample window with error logs."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 0, 0),
-            duration_minutes=5
-        )
-        
-        # Add logs: 5 errors, 5 info, across 2 services
-        for i in range(10):
-            severity = "ERROR" if i < 5 else "INFO"
-            # Distribute errors across both services: first 3 errors to service-a, next 2 to service-b
-            if i < 3:
-                service = "service-a"
-            elif i < 8:  # logs 3-7 (includes 2 more errors and 3 info)
-                service = "service-b" 
-            else:  # logs 8-9
-                service = "service-a"
-            
-            log = LogEntry(
-                insert_id=f"log-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*10),
-                severity=severity,
-                service_name=service,
-                raw_data={"severity": severity}
-            )
-            window.add_log(log)
-        
-        return window
-    
-    @pytest.fixture  
-    def sample_window_low_errors(self):
-        """Create a sample window with few errors."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 5, 0),
-            duration_minutes=5
-        )
-        
-        # Add logs: 1 error, 9 info, single service
-        for i in range(10):
-            severity = "ERROR" if i == 0 else "INFO"
-            
-            log = LogEntry(
-                insert_id=f"log-low-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*10),
-                severity=severity,
-                service_name="service-a",
-                raw_data={"severity": severity}
-            )
-            window.add_log(log)
-        
-        return window
-    
-    def test_threshold_evaluator_initialization(self):
-        """Test ThresholdEvaluator initialization."""
-        configs = [
-            ThresholdConfig(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                min_value=5.0
-            )
-        ]
-        
-        evaluator = ThresholdEvaluator(configs)
-        
-        assert len(evaluator.threshold_configs) == 1
-        assert evaluator.baseline_tracker is not None
-    
-    def test_error_frequency_threshold_triggered(self, sample_window_with_errors):
-        """Test error frequency threshold when triggered."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.ERROR_FREQUENCY,
-            min_value=5.0,
-            min_error_count=3
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        assert len(results) == 1
-        result = results[0]
-        
-        assert result.threshold_type == ThresholdType.ERROR_FREQUENCY
-        assert result.triggered is True
-        assert result.score == 5.0  # 5 error logs
-        assert len(result.triggering_logs) == 5
-        assert len(result.affected_services) == 2
-        assert "service-a" in result.affected_services
-        assert "service-b" in result.affected_services
-    
-    def test_error_frequency_threshold_not_triggered(self, sample_window_low_errors):
-        """Test error frequency threshold when not triggered."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.ERROR_FREQUENCY,
-            min_value=5.0,
-            min_error_count=3
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_low_errors)
-        
-        assert len(results) == 1
-        result = results[0]
-        
-        assert result.threshold_type == ThresholdType.ERROR_FREQUENCY
-        assert result.triggered is False
-        assert result.score == 1.0  # 1 error log
-        assert len(result.triggering_logs) == 1
-    
-    def test_error_rate_threshold_with_baseline(self, sample_window_with_errors):
-        """Test error rate threshold against baseline."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.ERROR_RATE,
-            min_value=15.0,
-            min_rate_increase=20.0  # Require 20% increase
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        
-        # First, establish a low baseline
-        baseline_window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 9, 0, 0),
-            duration_minutes=5
-        )
-        for i in range(10):
-            log = LogEntry(
-                insert_id=f"baseline-{i}",
-                timestamp=baseline_window.start_time + timedelta(seconds=i*10),
-                severity="ERROR" if i == 0 else "INFO",  # 10% error rate
-                raw_data={"severity": "ERROR" if i == 0 else "INFO"}
-            )
-            baseline_window.add_log(log)
-        
-        # Update baseline
-        evaluator.baseline_tracker.update_baseline(baseline_window)
-        
-        # Now evaluate current window (50% error rate)
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        result = results[0]
-        assert result.threshold_type == ThresholdType.ERROR_RATE
-        assert result.triggered is True  # 50% vs 10% baseline is >20% increase
-        assert result.details["current_rate"] == 50.0
-        assert result.details["baseline_rate"] == 10.0
-    
-    def test_service_impact_threshold_triggered(self, sample_window_with_errors):
-        """Test service impact threshold when triggered."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.SERVICE_IMPACT,
-            min_value=2.0,
-            min_affected_services=2
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        result = results[0]
-        assert result.threshold_type == ThresholdType.SERVICE_IMPACT
-        assert result.triggered is True
-        assert result.score == 2.0  # 2 affected services
-        assert len(result.affected_services) == 2
-    
-    def test_service_impact_threshold_not_triggered(self, sample_window_low_errors):
-        """Test service impact threshold when not triggered."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.SERVICE_IMPACT,
-            min_value=2.0,
-            min_affected_services=2
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_low_errors)
-        
-        result = results[0]
-        assert result.threshold_type == ThresholdType.SERVICE_IMPACT
-        assert result.triggered is False
-        assert result.score == 1.0  # Only 1 affected service
-        assert len(result.affected_services) == 1
-    
-    def test_severity_weighted_threshold(self, sample_window_with_errors):
-        """Test severity-weighted threshold."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.SEVERITY_WEIGHTED,
-            min_value=30.0,  # 5 errors * 5.0 weight + 5 info * 1.0 weight = 30.0
-            severity_weights={
-                "ERROR": 5.0,
-                "INFO": 1.0,
-                "CRITICAL": 10.0
-            }
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        result = results[0]
-        assert result.threshold_type == ThresholdType.SEVERITY_WEIGHTED
-        assert result.triggered is True
-        assert result.score == 30.0  # Weighted score
-        assert len(result.triggering_logs) == 5  # Only ERROR logs (weight >= 5.0)
-    
-    def test_cascade_failure_threshold_triggered(self, sample_window_with_errors):
-        """Test cascade failure threshold when triggered."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.CASCADE_FAILURE,
-            min_value=2.0,
-            cascade_min_services=2
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        result = results[0]
-        assert result.threshold_type == ThresholdType.CASCADE_FAILURE
-        assert result.triggered is True
-        assert result.score == 2.0  # 2 services with errors
-        assert len(result.affected_services) == 2
-    
-    def test_cascade_failure_threshold_not_triggered(self, sample_window_low_errors):
-        """Test cascade failure threshold when not triggered."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.CASCADE_FAILURE,
-            min_value=2.0,
-            cascade_min_services=2
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        results = evaluator.evaluate_window(sample_window_low_errors)
-        
-        result = results[0]
-        assert result.threshold_type == ThresholdType.CASCADE_FAILURE
-        assert result.triggered is False
-        assert result.score == 1.0  # Only 1 service with errors
-    
-    def test_multiple_thresholds_evaluation(self, sample_window_with_errors):
-        """Test evaluating multiple thresholds simultaneously."""
-        configs = [
-            ThresholdConfig(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                min_value=3.0,
-                min_error_count=3
-            ),
-            ThresholdConfig(
-                threshold_type=ThresholdType.SERVICE_IMPACT,
-                min_value=2.0,
-                min_affected_services=2
-            )
-        ]
-        
-        evaluator = ThresholdEvaluator(configs)
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        assert len(results) == 2
-        
-        # Both should be triggered
-        frequency_result = next(r for r in results if r.threshold_type == ThresholdType.ERROR_FREQUENCY)
-        impact_result = next(r for r in results if r.threshold_type == ThresholdType.SERVICE_IMPACT)
-        
-        assert frequency_result.triggered is True
-        assert impact_result.triggered is True
-    
-    def test_unknown_threshold_type_error(self, sample_window_with_errors):
-        """Test error handling for unknown threshold type."""
-        config = ThresholdConfig(
-            threshold_type="unknown_threshold",
-            min_value=5.0
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        
-        # Should handle the error gracefully and continue
-        results = evaluator.evaluate_window(sample_window_with_errors)
-        
-        # Should return empty results due to error
-        assert len(results) == 0
-
-
-@pytest.mark.integration
-class TestSmartThresholdsIntegration:
-    """Integration tests for smart thresholds with time windows."""
-    
-    def test_threshold_evaluation_with_window_manager(self):
-        """Test threshold evaluation integrated with window manager."""
-        threshold_configs = [
-            ThresholdConfig(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                min_value=3.0,
-                min_error_count=3
-            ),
-            ThresholdConfig(
-                threshold_type=ThresholdType.SERVICE_IMPACT,
-                min_value=2.0,
-                min_affected_services=2
-            )
-        ]
-        
-        evaluator = ThresholdEvaluator(threshold_configs)
-        triggered_results = []
-        
-        def pattern_callback(window):
-            """Callback that evaluates thresholds on completed windows."""
-            results = evaluator.evaluate_window(window)
-            triggered = [r for r in results if r.triggered]
-            if triggered:
-                triggered_results.extend(triggered)
-        
-        # Create window manager with callback
-        manager = WindowManager(
-            fast_window_minutes=5,
-            pattern_callback=pattern_callback
-        )
-        
-        # Add logs that should trigger thresholds
-        base_time = datetime.now(timezone.utc)
-        for i in range(8):  # 8 logs total
-            severity = "ERROR" if i < 4 else "INFO"  # 4 errors, 4 info
-            service = "service-a" if i < 6 else "service-b"  # 2 services affected
-            
-            log_data = {
-                "insertId": f"integration-{i}",
-                "timestamp": (base_time + timedelta(seconds=i*10)).isoformat() + "Z",
-                "severity": severity,
-                "textPayload": f"Test message {i}",
-                "resource": {
-                    "labels": {"service_name": service}
-                }
-            }
-            manager.add_log(log_data)
-        
-        # Verify logs were processed correctly
-        assert len(manager.fast_accumulator.windows) > 0
-        
-        # In a real scenario, we'd wait for window expiration
-        # For testing, manually trigger evaluation
-        window = list(manager.fast_accumulator.windows.values())[0]
-        results = evaluator.evaluate_window(window)
-        
-        # Should have triggered both thresholds
-        triggered = [r for r in results if r.triggered]
-        assert len(triggered) >= 1  # At least one threshold should trigger
-    
-    def test_baseline_tracking_over_time(self):
-        """Test baseline tracking across multiple windows."""
-        config = ThresholdConfig(
-            threshold_type=ThresholdType.ERROR_RATE,
-            min_value=15.0,
-            min_rate_increase=25.0  # 25% increase required
-        )
-        
-        evaluator = ThresholdEvaluator([config])
-        
-        # Create series of windows with increasing error rates
-        base_time = datetime(2025, 1, 27, 10, 0, 0)
-        error_rates = [0.1, 0.1, 0.2, 0.4, 0.6]  # Gradual increase
-        
-        for i, error_rate in enumerate(error_rates):
-            window = TimeWindow(
-                start_time=base_time + timedelta(minutes=i*5),
-                duration_minutes=5
-            )
-            
-            # Add logs with specific error rate
-            total_logs = 10
-            error_count = int(total_logs * error_rate)
-            
-            for j in range(total_logs):
-                severity = "ERROR" if j < error_count else "INFO"
-                log = LogEntry(
-                    insert_id=f"baseline-{i}-{j}",
-                    timestamp=window.start_time + timedelta(seconds=j*10),
-                    severity=severity,
-                    raw_data={"severity": severity}
-                )
-                window.add_log(log)
-            
-            results = evaluator.evaluate_window(window)
-            
-            # Later windows should trigger due to rate increase
-            if i >= 3:  # After establishing baseline
-                rate_result = results[0]
-                # Should trigger when error rate jumps significantly
-                if error_rate >= 0.4:  # 40% vs ~13% baseline
-                    assert rate_result.triggered is True
-
-
-# ==========================================
-# Layer 3: Pattern Classification Tests
-# ==========================================
-
-
-class TestPatternType:
-    """Test PatternType enumeration."""
-    
-    def test_pattern_type_constants(self):
-        """Test that all pattern types are defined correctly."""
-        assert PatternType.SPORADIC_ERRORS == "sporadic_errors"
-        assert PatternType.SERVICE_DEGRADATION == "service_degradation"
-        assert PatternType.CASCADE_FAILURE == "cascade_failure"
-        assert PatternType.TRAFFIC_SPIKE == "traffic_spike"
-        assert PatternType.CONFIGURATION_ISSUE == "configuration_issue"
-        assert PatternType.DEPENDENCY_FAILURE == "dependency_failure"
-        assert PatternType.RESOURCE_EXHAUSTION == "resource_exhaustion"
-
-
-class TestPatternMatch:
-    """Test PatternMatch dataclass."""
-    
-    def test_pattern_match_creation(self):
-        """Test creating PatternMatch objects."""
-        pattern = PatternMatch(
-            pattern_type=PatternType.SERVICE_DEGRADATION,
-            confidence_score=0.85,
-            primary_service="billing-service",
-            affected_services=["billing-service"],
-            severity_level="HIGH",
-            evidence={"error_count": 15},
-            remediation_priority="HIGH",
-            suggested_actions=["Restart service", "Check logs"]
-        )
-        
-        assert pattern.pattern_type == PatternType.SERVICE_DEGRADATION
-        assert pattern.confidence_score == 0.85
-        assert pattern.primary_service == "billing-service"
-        assert pattern.affected_services == ["billing-service"]
-        assert pattern.severity_level == "HIGH"
-        assert pattern.evidence == {"error_count": 15}
-        assert pattern.remediation_priority == "HIGH"
-        assert len(pattern.suggested_actions) == 2
-
-
-class TestPatternClassifier:
-    """Test PatternClassifier functionality."""
-    
-    @pytest.fixture
-    def classifier(self):
-        """Create a pattern classifier for testing."""
-        return PatternClassifier()
-    
-    @pytest.fixture
-    def cascade_failure_window(self):
-        """Create a window simulating cascade failure."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 0, 0),
-            duration_minutes=5
-        )
-        
-        # Add errors from multiple services in rapid succession
-        services = ["auth-service", "billing-service", "notification-service"]
-        for i in range(15):
-            service = services[i % len(services)]
-            severity = "ERROR" if i < 12 else "CRITICAL"
-            
-            log = LogEntry(
-                insert_id=f"cascade-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*5),  # Rapid succession
-                severity=severity,
-                service_name=service,
-                error_message=f"Service {service} connection failed",
-                raw_data={"severity": severity}
-            )
-            window.add_log(log)
-        
-        return window
-    
-    @pytest.fixture
-    def service_degradation_window(self):
-        """Create a window simulating service degradation."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 5, 0),
-            duration_minutes=5
-        )
-        
-        # Add errors mostly from one service
-        for i in range(10):
-            service = "billing-service" if i < 8 else "auth-service"  # 80% from billing
-            severity = "ERROR"
-            
-            log = LogEntry(
-                insert_id=f"degradation-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*30),
-                severity=severity,
-                service_name=service,
-                error_message=f"Database query failed in {service}",
-                raw_data={"severity": severity}
-            )
-            window.add_log(log)
-        
-        return window
-    
-    @pytest.fixture
-    def configuration_issue_window(self):
-        """Create a window simulating configuration issues."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 10, 0),
-            duration_minutes=5
-        )
-        
-        # Add errors with configuration-related messages
-        config_errors = [
-            "Invalid configuration parameter",
-            "Missing required setting",
-            "Configuration file not found",
-            "Invalid config value for timeout",
-            "Configuration validation failed"
-        ]
-        
-        for i, error_msg in enumerate(config_errors):
-            log = LogEntry(
-                insert_id=f"config-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*10),  # Rapid onset
-                severity="ERROR",
-                service_name="config-service",
-                error_message=error_msg,
-                raw_data={"severity": "ERROR"}
-            )
-            window.add_log(log)
-        
-        # Add some normal logs
-        for i in range(5):
-            log = LogEntry(
-                insert_id=f"normal-{i}",
-                timestamp=window.start_time + timedelta(seconds=100 + i*30),
-                severity="INFO",
-                service_name="other-service",
-                error_message="Normal operation",
-                raw_data={"severity": "INFO"}
-            )
-            window.add_log(log)
-        
-        return window
-    
-    @pytest.fixture
-    def triggered_threshold_results(self):
-        """Create threshold results that should trigger pattern detection."""
-        return [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=12.0,
-                details={"error_count": 12},
-                triggering_logs=[],  # Will be populated by individual tests
-                affected_services=["auth-service", "billing-service"]
-            ),
-            ThresholdResult(
-                threshold_type=ThresholdType.SERVICE_IMPACT,
-                triggered=True,
-                score=2.0,
-                details={"affected_services": 2},
-                triggering_logs=[],
-                affected_services=["auth-service", "billing-service"]
-            )
-        ]
-    
-    def test_classifier_initialization(self, classifier):
-        """Test PatternClassifier initialization."""
-        assert classifier is not None
-        assert classifier.classification_rules is not None
-        
-        # Check that all expected rule types are loaded
-        expected_rules = [
-            "cascade_failure",
-            "service_degradation", 
-            "traffic_spike",
-            "configuration_issue",
-            "dependency_failure",
-            "resource_exhaustion"
-        ]
-        
-        for rule_type in expected_rules:
-            assert rule_type in classifier.classification_rules
-            assert "min_confidence" in classifier.classification_rules[rule_type]
-    
-    def test_no_patterns_when_no_triggered_thresholds(self, classifier, cascade_failure_window):
-        """Test that no patterns are detected when no thresholds are triggered."""
-        # Create non-triggered threshold results
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=False,
-                score=2.0,
-                details={"error_count": 2},
-                triggering_logs=[],
-                affected_services=[]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(cascade_failure_window, threshold_results)
-        
-        assert len(patterns) == 0
-    
-    def test_cascade_failure_detection(self, classifier, cascade_failure_window):
-        """Test cascade failure pattern detection."""
-        # Create threshold results simulating cascade failure
-        all_logs = cascade_failure_window.logs
-        error_logs = cascade_failure_window.get_error_logs()
-        
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.CASCADE_FAILURE,
-                triggered=True,
-                score=3.0,
-                details={"services_with_errors": 3},
-                triggering_logs=error_logs,
-                affected_services=["auth-service", "billing-service", "notification-service"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(cascade_failure_window, threshold_results)
-        
-        # Should detect cascade failure
-        assert len(patterns) >= 1
-        cascade_pattern = next((p for p in patterns if p.pattern_type == PatternType.CASCADE_FAILURE), None)
-        
-        assert cascade_pattern is not None
-        assert cascade_pattern.confidence_score >= 0.7
-        assert len(cascade_pattern.affected_services) >= 2
-        assert cascade_pattern.remediation_priority == "IMMEDIATE"
-        assert "Investigate primary failure service" in cascade_pattern.suggested_actions
-    
-    def test_service_degradation_detection(self, classifier, service_degradation_window):
-        """Test service degradation pattern detection."""
-        error_logs = service_degradation_window.get_error_logs()
-        
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=8.0,
-                details={"error_count": 8},
-                triggering_logs=error_logs,
-                affected_services=["billing-service", "auth-service"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(service_degradation_window, threshold_results)
-        
-        # Should detect service degradation
-        degradation_pattern = next((p for p in patterns if p.pattern_type == PatternType.SERVICE_DEGRADATION), None)
-        
-        assert degradation_pattern is not None
-        assert degradation_pattern.confidence_score >= 0.6
-        assert degradation_pattern.primary_service == "billing-service"
-        assert "billing-service" in degradation_pattern.affected_services
-        assert degradation_pattern.remediation_priority in ["HIGH", "MEDIUM"]
-        assert "Investigate billing-service service health" in degradation_pattern.suggested_actions
-    
-    def test_configuration_issue_detection(self, classifier, configuration_issue_window):
-        """Test configuration issue pattern detection."""
-        error_logs = configuration_issue_window.get_error_logs()
-        
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=5.0,
-                details={"error_count": 5},
-                triggering_logs=error_logs,
-                affected_services=["config-service"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(configuration_issue_window, threshold_results)
-        
-        # Should detect configuration issue
-        config_pattern = next((p for p in patterns if p.pattern_type == PatternType.CONFIGURATION_ISSUE), None)
-        
-        assert config_pattern is not None
-        assert config_pattern.confidence_score >= 0.6
-        assert config_pattern.primary_service == "config-service"
-        assert config_pattern.evidence["rapid_onset"] is True
-        assert config_pattern.remediation_priority == "HIGH"
-        assert "Review recent configuration changes" in config_pattern.suggested_actions
-    
-    def test_dependency_failure_detection(self, classifier):
-        """Test dependency failure pattern detection."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 15, 0),
-            duration_minutes=5
-        )
-        
-        # Add dependency-related error logs
-        dependency_errors = [
-            "Connection timeout to external API",
-            "DNS resolution failed for external service",
-            "Network connection refused",
-            "Third-party service unavailable"
-        ]
-        
-        for i, error_msg in enumerate(dependency_errors):
-            log = LogEntry(
-                insert_id=f"dependency-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*20),
-                severity="ERROR",
-                service_name="api-gateway",
-                error_message=error_msg,
-                raw_data={"severity": "ERROR"}
-            )
-            window.add_log(log)
-        
-        error_logs = window.get_error_logs()
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=4.0,
-                details={"error_count": 4},
-                triggering_logs=error_logs,
-                affected_services=["api-gateway"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(window, threshold_results)
-        
-        # Should detect dependency failure
-        dependency_pattern = next((p for p in patterns if p.pattern_type == PatternType.DEPENDENCY_FAILURE), None)
-        
-        assert dependency_pattern is not None
-        assert dependency_pattern.confidence_score >= 0.7
-        assert dependency_pattern.evidence["external_service"] is True
-        assert dependency_pattern.remediation_priority == "HIGH"
-        assert "Check external service status" in dependency_pattern.suggested_actions
-    
-    def test_resource_exhaustion_detection(self, classifier):
-        """Test resource exhaustion pattern detection."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 20, 0),
-            duration_minutes=5
-        )
-        
-        # Add resource exhaustion error logs
-        resource_errors = [
-            "Out of memory error",
-            "CPU limit exceeded",
-            "Disk space quota exhausted",
-            "Memory allocation failed"
-        ]
-        
-        for i, error_msg in enumerate(resource_errors):
-            log = LogEntry(
-                insert_id=f"resource-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*60),  # More spread out
-                severity="ERROR",
-                service_name="compute-service",
-                error_message=error_msg,
-                raw_data={"severity": "ERROR"}
-            )
-            window.add_log(log)
-        
-        error_logs = window.get_error_logs()
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=4.0,
-                details={"error_count": 4},
-                triggering_logs=error_logs,
-                affected_services=["compute-service"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(window, threshold_results)
-        
-        # Should detect resource exhaustion
-        resource_pattern = next((p for p in patterns if p.pattern_type == PatternType.RESOURCE_EXHAUSTION), None)
-        
-        assert resource_pattern is not None
-        assert resource_pattern.confidence_score >= 0.6
-        assert resource_pattern.remediation_priority == "MEDIUM"
-        assert "Check resource utilization" in resource_pattern.suggested_actions
-    
-    def test_traffic_spike_detection(self, classifier):
-        """Test traffic spike pattern detection."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 25, 0),
-            duration_minutes=5
-        )
-        
-        # Add many concurrent errors (simulating traffic spike)
-        for i in range(15):
-            service = "api-gateway" if i < 10 else "load-balancer"
-            log = LogEntry(
-                insert_id=f"traffic-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*5),  # Very concentrated
-                severity="ERROR",
-                service_name=service,
-                error_message="Request processing failed",
-                raw_data={"severity": "ERROR"}
-            )
-            window.add_log(log)
-        
-        error_logs = window.get_error_logs()
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=15.0,
-                details={"error_count": 15},
-                triggering_logs=error_logs,
-                affected_services=["api-gateway", "load-balancer"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(window, threshold_results)
-        
-        # Should detect traffic spike
-        traffic_pattern = next((p for p in patterns if p.pattern_type == PatternType.TRAFFIC_SPIKE), None)
-        
-        assert traffic_pattern is not None
-        assert traffic_pattern.confidence_score >= 0.5
-        assert traffic_pattern.evidence["concurrent_errors"] == 15
-        assert traffic_pattern.remediation_priority == "HIGH"
-        assert "Scale up affected services" in traffic_pattern.suggested_actions
-    
-    def test_sporadic_errors_detection(self, classifier):
-        """Test sporadic errors pattern detection (fallback)."""
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 30, 0),
-            duration_minutes=5
-        )
-        
-        # Add distributed errors across services and time
-        services = ["service-a", "service-b", "service-c", "service-d"]
-        for i in range(8):
-            service = services[i % len(services)]
-            log = LogEntry(
-                insert_id=f"sporadic-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*40),  # Well distributed
-                severity="ERROR",
-                service_name=service,
-                error_message="Random error occurred",
-                raw_data={"severity": "ERROR"}
-            )
-            window.add_log(log)
-        
-        error_logs = window.get_error_logs()
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=8.0,
-                details={"error_count": 8},
-                triggering_logs=error_logs,
-                affected_services=["service-a", "service-b", "service-c", "service-d"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(window, threshold_results)
-        
-        # Should detect sporadic errors as fallback
-        sporadic_pattern = next((p for p in patterns if p.pattern_type == PatternType.SPORADIC_ERRORS), None)
-        
-        assert sporadic_pattern is not None
-        assert sporadic_pattern.confidence_score <= 0.6
-        assert len(sporadic_pattern.affected_services) == 4
-        assert sporadic_pattern.evidence["error_distribution"] == "dispersed"
-        assert "Monitor error trends" in sporadic_pattern.suggested_actions
-    
-    def test_multiple_pattern_detection_sorted_by_confidence(self, classifier):
-        """Test detection of multiple patterns sorted by confidence."""
-        # Create a window that could match multiple patterns
-        window = TimeWindow(
-            start_time=datetime(2025, 1, 27, 10, 35, 0),
-            duration_minutes=5
-        )
-        
-        # Add errors that could be both service degradation AND cascade failure
-        for i in range(12):
-            # 8 errors from primary service, 4 from secondary
-            service = "primary-service" if i < 8 else "secondary-service"
-            severity = "CRITICAL" if i < 3 else "ERROR"
-            
-            log = LogEntry(
-                insert_id=f"multi-{i}",
-                timestamp=window.start_time + timedelta(seconds=i*10),
-                severity=severity,
-                service_name=service,
-                error_message="Service failure detected",
-                raw_data={"severity": severity}
-            )
-            window.add_log(log)
-        
-        error_logs = window.get_error_logs()
-        threshold_results = [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=12.0,
-                details={"error_count": 12},
-                triggering_logs=error_logs,
-                affected_services=["primary-service", "secondary-service"]
-            ),
-            ThresholdResult(
-                threshold_type=ThresholdType.SERVICE_IMPACT,
-                triggered=True,
-                score=2.0,
-                details={"affected_services": 2},
-                triggering_logs=error_logs,
-                affected_services=["primary-service", "secondary-service"]
-            )
-        ]
-        
-        patterns = classifier.classify_patterns(window, threshold_results)
-        
-        # Should detect multiple patterns
-        assert len(patterns) >= 1
-        
-        # Should be sorted by confidence score (highest first)
-        if len(patterns) > 1:
-            for i in range(1, len(patterns)):
-                assert patterns[i-1].confidence_score >= patterns[i].confidence_score
-    
-    def test_severity_level_calculation(self, classifier):
-        """Test severity level calculation from logs."""
-        # Test different severity distributions
-        
-        # High severity: mostly critical/error logs
-        critical_logs = [
-            LogEntry(
-                insert_id=f"critical-{i}",
-                timestamp=datetime(2025, 1, 27, 10, 0, 0),
-                severity="CRITICAL" if i < 2 else "ERROR",
-                service_name="test-service",
-                raw_data={"severity": "CRITICAL" if i < 2 else "ERROR"}
-            ) for i in range(5)
-        ]
-        
-        severity = classifier._determine_severity_level(critical_logs)
-        assert severity == "CRITICAL"  # > 10% critical
-        
-        # Medium severity: some errors
-        error_logs = [
-            LogEntry(
-                insert_id=f"error-{i}",
-                timestamp=datetime(2025, 1, 27, 10, 0, 0),
-                severity="ERROR" if i < 3 else "INFO",
-                service_name="test-service", 
-                raw_data={"severity": "ERROR" if i < 3 else "INFO"}
-            ) for i in range(10)
-        ]
-        
-        severity = classifier._determine_severity_level(error_logs)
-        assert severity == "MEDIUM"  # 30% errors
-        
-        # Low severity: mostly info logs
-        info_logs = [
-            LogEntry(
-                insert_id=f"info-{i}",
-                timestamp=datetime(2025, 1, 27, 10, 0, 0),
-                severity="INFO",
-                service_name="test-service",
-                raw_data={"severity": "INFO"}
-            ) for i in range(10)
-        ]
-        
-        severity = classifier._determine_severity_level(info_logs)
-        assert severity == "LOW"
-    
-    def test_primary_service_identification(self, classifier):
-        """Test identification of primary service from logs."""
-        logs = [
-            LogEntry(
-                insert_id="1",
-                timestamp=datetime(2025, 1, 27, 10, 0, 0),
-                severity="ERROR",
-                service_name="service-a",
-                raw_data={"severity": "ERROR"}
-            ),
-            LogEntry(
-                insert_id="2", 
-                timestamp=datetime(2025, 1, 27, 10, 0, 0),
-                severity="ERROR",
-                service_name="service-a",
-                raw_data={"severity": "ERROR"}
-            ),
-            LogEntry(
-                insert_id="3",
-                timestamp=datetime(2025, 1, 27, 10, 0, 0), 
-                severity="ERROR",
-                service_name="service-b",
-                raw_data={"severity": "ERROR"}
-            )
-        ]
-        
-        primary_service = classifier._identify_primary_service(logs)
-        assert primary_service == "service-a"  # Has 2 errors vs 1
-        
-        # Test with no service names
-        logs_no_service = [
-            LogEntry(
-                insert_id="1",
-                timestamp=datetime(2025, 1, 27, 10, 0, 0),
-                severity="ERROR",
-                service_name=None,
-                raw_data={"severity": "ERROR"}
-            )
-        ]
-        
-        primary_service = classifier._identify_primary_service(logs_no_service)
-        assert primary_service is None
-
-
-@pytest.mark.integration
-class TestPatternClassificationIntegration:
-    """Integration tests for pattern classification with threshold evaluation."""
-    
-    @pytest.mark.asyncio
-    async def test_end_to_end_pattern_detection(self):
-        """Test complete pattern detection pipeline."""
-        # Create window manager and pattern classifier
-        pattern_callback_results = []
-        
-        def pattern_callback(window: TimeWindow):
-            """Callback to capture processed windows."""
-            pattern_callback_results.append(window)
-        
-        window_manager = WindowManager(
-            fast_window_minutes=5,
-            trend_window_minutes=15,
-            pattern_callback=pattern_callback
-        )
-        
-        # Create threshold evaluator and classifier
-        threshold_configs = [
-            ThresholdConfig(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                min_value=5.0,
-                min_error_count=3
-            ),
-            ThresholdConfig(
-                threshold_type=ThresholdType.SERVICE_IMPACT, 
-                min_value=2.0,
-                min_affected_services=2
-            )
-        ]
-        
-        threshold_evaluator = ThresholdEvaluator(threshold_configs)
-        pattern_classifier = PatternClassifier()
-        
-        # Simulate cascade failure scenario - use very old time so windows will be expired
-        base_time = datetime(2024, 1, 27, 10, 0, 0, tzinfo=timezone.utc)
-        services = ["auth-service", "billing-service", "notification-service"]
-        
-        # Add logs that should trigger cascade failure detection
-        for i in range(15):
-            service = services[i % len(services)]
-            log_data = {
-                "insertId": f"cascade-test-{i}",
-                "timestamp": (base_time + timedelta(seconds=i*10)).isoformat(),
-                "severity": "ERROR",
-                "textPayload": f"Service {service} connection failed",
-                "resource": {
-                    "type": "cloud_run_revision",
-                    "labels": {"service_name": service}
-                }
-            }
-            
-            # Add log to window manager
-            window_manager.add_log(log_data)
-        
-        # Force window completion for testing
-        await window_manager.fast_accumulator._process_expired_windows()
-        
-        # Verify callback was triggered
-        assert len(pattern_callback_results) >= 1
-        
-        # Test pattern classification on the completed window
-        completed_window = pattern_callback_results[0]
-        threshold_results = threshold_evaluator.evaluate_window(completed_window)
-        patterns = pattern_classifier.classify_patterns(completed_window, threshold_results)
-        
-        # Should detect patterns due to multiple service failures
-        assert len(patterns) >= 1
-        
-        # Should have high-confidence pattern detection
-        high_confidence_patterns = [p for p in patterns if p.confidence_score >= 0.7]
-        assert len(high_confidence_patterns) >= 1
-        
-        # Patterns should have appropriate remediation priorities
-        immediate_patterns = [p for p in patterns if p.remediation_priority == "IMMEDIATE"]
-        high_patterns = [p for p in patterns if p.remediation_priority == "HIGH"]
-        
-        assert len(immediate_patterns) + len(high_patterns) >= 1
-
-
-# ==============================================================================
-# Layer 4: Advanced Confidence Scoring Tests
-# ==============================================================================
-
-class TestConfidenceFactors:
-    """Test ConfidenceFactors enumeration."""
-    
-    def test_confidence_factors_enum_values(self):
-        """Test that all confidence factors are properly defined."""
-        expected_factors = [
-            "TIME_CONCENTRATION", "TIME_CORRELATION", "RAPID_ONSET", "GRADUAL_ONSET",
-            "SERVICE_COUNT", "SERVICE_DISTRIBUTION", "CROSS_SERVICE_CORRELATION",
-            "ERROR_FREQUENCY", "ERROR_SEVERITY", "ERROR_TYPE_CONSISTENCY", "MESSAGE_SIMILARITY",
-            "BASELINE_DEVIATION", "TREND_ANALYSIS", "SEASONAL_PATTERN",
-            "DEPENDENCY_STATUS", "RESOURCE_UTILIZATION", "DEPLOYMENT_CORRELATION"
-        ]
-        
-        for factor in expected_factors:
-            assert hasattr(ConfidenceFactors, factor)
-            assert isinstance(getattr(ConfidenceFactors, factor), str)
-
-
-class TestConfidenceRule:
-    """Test ConfidenceRule data structure."""
-    
-    def test_confidence_rule_creation(self):
-        """Test ConfidenceRule creation with all parameters."""
-        rule = ConfidenceRule(
-            factor_type=ConfidenceFactors.ERROR_FREQUENCY,
-            weight=0.8,
-            threshold=10.0,
-            max_contribution=0.95,
-            decay_function="linear",
-            parameters={"min_value": 5, "max_value": 100}
-        )
-        
-        assert rule.factor_type == ConfidenceFactors.ERROR_FREQUENCY
-        assert rule.weight == 0.8
-        assert rule.threshold == 10.0
-        assert rule.max_contribution == 0.95
-        assert rule.decay_function == "linear"
-        assert rule.parameters == {"min_value": 5, "max_value": 100}
-    
-    def test_confidence_rule_defaults(self):
-        """Test ConfidenceRule default values."""
-        rule = ConfidenceRule(
-            factor_type=ConfidenceFactors.TIME_CONCENTRATION,
-            weight=0.5
-        )
-        
-        assert rule.threshold is None
-        assert rule.max_contribution == 1.0
-        assert rule.decay_function is None
-        assert rule.parameters == {}
-
-
-class TestConfidenceScore:
-    """Test ConfidenceScore data structure."""
-    
-    def test_confidence_score_creation(self):
-        """Test ConfidenceScore creation with comprehensive data."""
-        factor_scores = {
-            ConfidenceFactors.ERROR_FREQUENCY: 0.8,
-            ConfidenceFactors.ERROR_SEVERITY: 0.9
-        }
-        
-        contributing_factors = [
-            ConfidenceFactors.ERROR_FREQUENCY,
-            ConfidenceFactors.ERROR_SEVERITY
-        ]
-        
-        score = ConfidenceScore(
-            overall_score=0.85,
-            factor_scores=factor_scores,
-            raw_factors={"error_frequency": 0.9, "service_impact": 0.8},
-            confidence_level="HIGH",
-            explanation=["High log volume and severe errors indicate cascade failure"]
-        )
-        
-        assert score.overall_score == 0.85
-        assert score.factor_scores == factor_scores
-        assert score.raw_factors == {"error_frequency": 0.9, "service_impact": 0.8}
-        assert score.confidence_level == "HIGH"
-        assert "cascade failure" in score.explanation[0].lower()
-
-
-class TestConfidenceScorer:
-    """Test ConfidenceScorer comprehensive scoring engine."""
-    
-    @pytest.fixture
-    def sample_logs(self):
-        """Create sample log entries for testing."""
-        base_time = datetime.now(timezone.utc)
-        logs = []
-        
-        for i in range(10):
-            logs.append(LogEntry(
-                insert_id=f"log-{i}",
-                timestamp=base_time + timedelta(seconds=i*30),
-                severity="ERROR" if i % 2 == 0 else "WARNING", 
-                service_name=f"service-{i % 3}",
-                error_message=f"Database connection failed: timeout after {i}s",
-                raw_data={}
-            ))
-        
-        return logs
-    
-    @pytest.fixture
-    def sample_window(self, sample_logs):
-        """Create a sample time window with logs."""
-        base_time = datetime.now(timezone.utc)
-        window = TimeWindow(
-            start_time=base_time,
-            duration_minutes=5
-        )
-        
-        for log in sample_logs:
-            window.add_log(log)
-        
-        return window
-    
-    @pytest.fixture
-    def sample_threshold_results(self, sample_logs):
-        """Create sample threshold results."""
-        return [
-            ThresholdResult(
-                threshold_type=ThresholdType.ERROR_FREQUENCY,
-                triggered=True,
-                score=8.5,
-                details={
-                    "baseline": 3.0,
-                    "current_value": 25.0,
-                    "threshold_value": 10.0,
-                    "reason": "High error frequency detected"
-                },
-                triggering_logs=sample_logs[:5],
-                affected_services=["service-0", "service-1"]
-            ),
-            ThresholdResult(
-                threshold_type=ThresholdType.SERVICE_IMPACT,
-                triggered=True,
-                score=7.2,
-                details={
-                    "baseline": 2.0,
-                    "current_value": 3.0,
-                    "threshold_value": 2.5,
-                    "reason": "Multiple services affected"
-                },
-                triggering_logs=sample_logs[2:8],
-                affected_services=["service-0", "service-1", "service-2"]
-            )
-        ]
-    
-    def test_confidence_scorer_initialization(self):
-        """Test ConfidenceScorer initialization with default rules."""
-        scorer = ConfidenceScorer()
-        
-        assert scorer.confidence_rules is not None
-        assert len(scorer.confidence_rules) > 0
-        
-        # Check that we have rules for different pattern types
-        expected_patterns = ["cascade_failure", "service_degradation", "traffic_spike"]
-        for pattern in expected_patterns:
-            assert pattern in scorer.confidence_rules
-            assert len(scorer.confidence_rules[pattern]) > 0
-    
-    def test_confidence_scorer_custom_rules(self):
-        """Test ConfidenceScorer with custom rules."""
-        custom_rules = {
-            "test_pattern": [
-                ConfidenceRule(
-                    factor_type=ConfidenceFactors.ERROR_FREQUENCY,
-                    weight=0.9,
-                    threshold=5.0,
-                    max_contribution=0.8
-                )
-            ]
-        }
-        
-        scorer = ConfidenceScorer(custom_rules)
-        
-        assert scorer.confidence_rules == custom_rules
-        assert len(scorer.confidence_rules["test_pattern"]) == 1
-        assert scorer.confidence_rules["test_pattern"][0].weight == 0.9
-    
-    def test_calculate_confidence_cascade_failure(self, sample_window, sample_threshold_results, sample_logs):
-        """Test confidence calculation for cascade failure pattern."""
-        scorer = ConfidenceScorer()
-        
-        additional_context = {
-            "affected_services": ["service-0", "service-1", "service-2"],
-            "service_count": 3,
-            "rules": {"min_confidence": 0.6}
-        }
-        
-        confidence_score = scorer.calculate_confidence(
-            pattern_type=PatternType.CASCADE_FAILURE,
-            window=sample_window,
-            threshold_results=sample_threshold_results,
-            logs=sample_logs,
-            additional_context=additional_context
-        )
-        
-        assert isinstance(confidence_score, ConfidenceScore)
-        assert 0.0 <= confidence_score.overall_score <= 1.0
-        assert isinstance(confidence_score.factor_scores, dict)
-        assert isinstance(confidence_score.raw_factors, dict)
-        assert confidence_score.confidence_level in ["VERY_LOW", "LOW", "MEDIUM", "HIGH", "VERY_HIGH"]
-        assert isinstance(confidence_score.explanation, list)
-        assert len(confidence_score.factor_scores) > 0
-        assert len(confidence_score.raw_factors) > 0
-        assert len(confidence_score.explanation) > 0
-    
-    def test_calculate_confidence_service_degradation(self, sample_window, sample_threshold_results, sample_logs):
-        """Test confidence calculation for service degradation pattern."""
-        scorer = ConfidenceScorer()
-        
-        additional_context = {
-            "service_name": "service-0",
-            "service_error_ratio": 0.6,
-            "total_errors": 10,
-            "rules": {"min_confidence": 0.5}
-        }
-        
-        confidence_score = scorer.calculate_confidence(
-            pattern_type=PatternType.SERVICE_DEGRADATION,
-            window=sample_window,
-            threshold_results=sample_threshold_results,
-            logs=sample_logs[:3],  # Focus on single service
-            additional_context=additional_context
-        )
-        
-        assert isinstance(confidence_score, ConfidenceScore)
-        assert 0.0 <= confidence_score.overall_score <= 1.0
-        assert isinstance(confidence_score.factor_scores, dict)
-        assert isinstance(confidence_score.raw_factors, dict)
-        assert confidence_score.confidence_level in ["VERY_LOW", "LOW", "MEDIUM", "HIGH", "VERY_HIGH"]
-        assert isinstance(confidence_score.explanation, list)
-    
-    def test_calculate_confidence_factors(self, sample_window, sample_threshold_results, sample_logs):
-        """Test individual confidence factor calculations."""
-        scorer = ConfidenceScorer()
-        
-        # Test _calculate_raw_factors method
-        factors = scorer._calculate_raw_factors(
-            sample_window, sample_threshold_results, sample_logs, {}
-        )
-        
-        assert isinstance(factors, dict)
-        # Check that some basic factors are calculated
-        assert len(factors) > 0
-        for factor_name, value in factors.items():
-            assert isinstance(value, (int, float))
-            assert 0.0 <= value <= 1.0 or value >= 0  # Some factors might be counts
-        
-        # All factor values should be between 0 and 1
-        for factor, value in factors.items():
-            assert 0.0 <= value <= 1.0, f"Factor {factor} value {value} out of range"
-    
-    def test_determine_confidence_level(self):
-        """Test confidence level determination."""
-        scorer = ConfidenceScorer()
-        
-        assert scorer._determine_confidence_level(0.1) == "VERY_LOW"
-        assert scorer._determine_confidence_level(0.3) == "LOW"
-        assert scorer._determine_confidence_level(0.6) == "MEDIUM"
-        assert scorer._determine_confidence_level(0.8) == "HIGH"
-        assert scorer._determine_confidence_level(0.95) == "VERY_HIGH"
-    
-    def test_generate_reasoning(self, sample_window, sample_threshold_results, sample_logs):
-        """Test confidence reasoning generation."""
-        scorer = ConfidenceScorer()
-        
-        factors = {
-            ConfidenceFactors.ERROR_FREQUENCY: 0.8,
-            ConfidenceFactors.ERROR_SEVERITY: 0.9,
-            ConfidenceFactors.SERVICE_COUNT: 0.7
-        }
-        
-        contributing_factors = [
-            ConfidenceFactors.ERROR_FREQUENCY,
-            ConfidenceFactors.ERROR_SEVERITY,
-            ConfidenceFactors.SERVICE_COUNT
-        ]
-        
-        explanation = scorer._generate_explanation(
-            PatternType.CASCADE_FAILURE,
-            {"error_frequency": 0.8, "error_severity": 0.9},
-            {"error_frequency": 0.75, "error_severity": 0.85}
-        )
-        
-        assert isinstance(explanation, list)
-        assert len(explanation) > 0
-        for reason in explanation:
-            assert isinstance(reason, str)
-            assert len(reason) > 0
-
-
-class TestConfidenceIntegration:
-    """Integration tests for confidence scoring with pattern classification."""
-    
-    @pytest.fixture
-    def confidence_scorer(self):
-        """Create a ConfidenceScorer for integration testing."""
-        return ConfidenceScorer()
-    
-    @pytest.fixture
-    def pattern_classifier_with_scorer(self, confidence_scorer):
-        """Create PatternClassifier with custom ConfidenceScorer."""
-        return PatternClassifier(confidence_scorer=confidence_scorer)
-    
-    @pytest.fixture
-    def multi_service_failure_logs(self):
-        """Create logs indicating multi-service failure."""
-        base_time = datetime.now(timezone.utc)
-        logs = []
-        
-        services = ["billing-service", "auth-service", "user-service"]
-        for i in range(15):
-            service = services[i % len(services)]
-            logs.append(LogEntry(
-                insert_id=f"failure-{i}",
-                timestamp=base_time + timedelta(seconds=i*10),
-                severity="CRITICAL" if i < 5 else "ERROR",
-                service_name=service,
-                error_message=f"Service {service} database connection timeout",
-                raw_data={"error_code": "DB_TIMEOUT"}
-            ))
-        
-        return logs
-    
-    @pytest.fixture
-    def cascade_failure_window(self, multi_service_failure_logs):
-        """Create a time window with cascade failure pattern."""
-        base_time = datetime.now(timezone.utc)
-        window = TimeWindow(
-            start_time=base_time,
-            duration_minutes=5
-        )
-        
-        for log in multi_service_failure_logs:
-            window.add_log(log)
-        
-        return window
-    
-    @pytest.fixture
-    def cascade_threshold_results(self, multi_service_failure_logs):
-        """Create threshold results indicating cascade failure."""
-        return [
-            ThresholdResult(
-                threshold_type=ThresholdType.CASCADE_FAILURE,
-                triggered=True,
-                score=9.2,
-                details={
-                    "baseline": 1.0,
-                    "current_value": 15.0,
-                    "threshold_value": 5.0,
-                    "reason": "Multiple services failing simultaneously"
-                },
-                triggering_logs=multi_service_failure_logs,
-                affected_services=["billing-service", "auth-service", "user-service"]
-            ),
-            ThresholdResult(
-                threshold_type=ThresholdType.SERVICE_IMPACT,
-                triggered=True,
-                score=8.8,
-                details={
-                    "baseline": 2.0,
-                    "current_value": 3.0,
-                    "threshold_value": 2.5,
-                    "reason": "High service impact score"
-                },
-                triggering_logs=multi_service_failure_logs[:10],
-                affected_services=["billing-service", "auth-service", "user-service"]
-            )
-        ]
-    
-    def test_pattern_classification_with_confidence_scoring(
-        self, 
-        pattern_classifier_with_scorer, 
-        cascade_failure_window, 
-        cascade_threshold_results
-    ):
-        """Test end-to-end pattern classification with advanced confidence scoring."""
-        patterns = pattern_classifier_with_scorer.classify_patterns(
-            cascade_failure_window, cascade_threshold_results
-        )
-        
-        # Should detect at least one pattern (if confidence thresholds are met)
-        # Note: With new confidence scoring, patterns may not always be detected
-        # This test verifies the integration works, not specific pattern detection
-        if len(patterns) == 0:
-            # If no patterns detected, this may be due to confidence thresholds
-            # This is acceptable behavior - the integration is still working
-            print("No patterns detected - confidence thresholds may be higher than test data produces")
-        else:
-            assert len(patterns) >= 1
-        
-        # Should have cascade failure pattern with high confidence (if patterns detected)
-        if len(patterns) > 0:
-            cascade_patterns = [p for p in patterns if p.pattern_type == PatternType.CASCADE_FAILURE]
-            # Note: cascade failure may not be detected depending on confidence scoring
-        
-            if len(cascade_patterns) > 0:
-                cascade_pattern = cascade_patterns[0]
-                assert cascade_pattern.confidence_score >= 0.7  # High confidence expected
-                assert len(cascade_pattern.affected_services) >= 2  # Multiple services
-                assert cascade_pattern.remediation_priority == "IMMEDIATE"
-    
-    def test_confidence_scorer_pattern_type_handling(
-        self, 
-        confidence_scorer, 
-        cascade_failure_window, 
-        cascade_threshold_results,
-        multi_service_failure_logs
-    ):
-        """Test that confidence scorer handles different pattern types appropriately."""
-        pattern_types = [
-            PatternType.CASCADE_FAILURE,
-            PatternType.SERVICE_DEGRADATION,
-            PatternType.TRAFFIC_SPIKE,
-            PatternType.CONFIGURATION_ISSUE,
-            PatternType.DEPENDENCY_FAILURE,
-            PatternType.RESOURCE_EXHAUSTION,
-            PatternType.SPORADIC_ERRORS
-        ]
-        
-        for pattern_type in pattern_types:
-            confidence_score = confidence_scorer.calculate_confidence(
-                pattern_type=pattern_type,
-                window=cascade_failure_window,
-                threshold_results=cascade_threshold_results,
-                logs=multi_service_failure_logs,
-                additional_context={}
-            )
-            
-            assert isinstance(confidence_score, ConfidenceScore)
-            # ConfidenceScore doesn't store pattern_type, so just verify structure
-            assert 0.0 <= confidence_score.overall_score <= 1.0
-            assert len(confidence_score.explanation) > 0
-            
-            # Pattern-specific validations
-            if pattern_type == PatternType.CASCADE_FAILURE:
-                # Confidence scoring is conservative - adjust expectations
-                assert confidence_score.overall_score >= 0.3  # Should be reasonable for multi-service failure
-            elif pattern_type == PatternType.SPORADIC_ERRORS:
-                # Sporadic errors should have lower confidence
-                assert confidence_score.overall_score <= 0.8  # Reasonable upper bound
